[SIMP-2545] puppetagent_cron.sh needs updating (#43)
The shell script makes assumptions based upon behavior in older versions of Puppet. For instance, the puppet client service no longer returns in a process listing for 'puppetd'.

Also, I've found that sometimes a Puppet agent can become unresponsive ("hung") and the script doesn't take that scenario into account. It will not run as long as it thinks another catalog run is in progress.