(SIMP-3280) Add SHA256-based random minute option to agent cron
Add SHA256-based option to generate the minute parameter for
a client's puppet agent cron from its IP address. This option
is intended mitigate the undesirable clustering of client puppet
agent runs, when the number of IPs to be transformed is less
than the minute range over which the randomization is requested
(60) and/or the client IPs are not linearly assigned.