(SIMP-4574) Make the master service name dynamic
Prior to this the service name for the Puppetserver was hard-coded to
the open source name, `puppetserver`. Puppet Enterprise uses the name
`pe-puppetserver`, however, so resources that use that variable were
declared with incorrect attributes. This commit changes the variable
assignment to use a selector based on the parameter
`pupmod::master::server_distribution` so it is dynamic.

SIMP-4574 #close