(SIMP-10390) Allow multiple puppet servers
Previously allowed a single Puppet server to be defined in puppet.conf
in the `server` setting.
This change allows for multiple servers to be set,
configuring server_list correctly if that is the case.
This change will allow larger organizations with multiple Puppet
servers to utilize the module.

SIMP-10390 #close